resolve function returning promise
